+++ pkg/util/scutil/coalescer.go
@@ -1,4 +1,4 @@
-package coalesce
+package scutil
 
 import (
 	"context"
